requirements.txt
python-dotenv==1.0.1
openai==1.42.0
pandas==2.2.3
